DEV Move s/p/collections to s/p/_collections
Totally forgot that collections is an actual python module
while writing this file. For some reason, python3 didn't
fail due to this aliasing (sensitivity reader trying to
import something from collections) by python2.7 did
